Add android mask strategy
